+++ test/assets/index.py
@@ -1,0 +1,29 @@
+import os.path
+
+TEST_DIR = os.path.dirname(__file__)
+
+TABLEAU_82_TWB = os.path.join(TEST_DIR, 'TABLEAU_82_TWB.twb')
+
+TABLEAU_93_TWB = os.path.join(TEST_DIR, 'TABLEAU_93_TWB.twb')
+
+TABLEAU_93_TDS = os.path.join(TEST_DIR, 'TABLEAU_93_TDS.tds')
+
+TABLEAU_10_TDS = os.path.join(TEST_DIR, 'TABLEAU_10_TDS.tds')
+
+TABLEAU_10_TWB = os.path.join(TEST_DIR, 'TABLEAU_10_TWB.twb')
+
+TABLEAU_CONNECTION_XML = os.path.join(TEST_DIR, 'CONNECTION.xml')
+
+TABLEAU_10_TWBX = os.path.join(TEST_DIR, 'TABLEAU_10_TWBX.twbx')
+
+TABLEAU_10_TDSX = os.path.join(TEST_DIR, 'TABLEAU_10_TDSX.tdsx')
+
+EMPTY_WORKBOOK = os.path.join(TEST_DIR, 'empty_workbook.twb')
+
+MULTI_CONNECTION_10 = os.path.join(TEST_DIR, 'multiple_connections.twb')
+
+BAD_ZIP_FILE = os.path.join(TEST_DIR, 'BadZip.zip')
+
+TWBX_WITH_CACHE_FILES = os.path.join(TEST_DIR, 'Cache.twbx')
+
+COMPLEX_TWB = os.path.join(TEST_DIR, 'filtering.twb')